git-svn-id: svn://192.168.1.10/public/CoffeeMud@9673 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Druid/Chant_CrystalGrowth.java
@@ -41,7 +41,7 @@ public class Chant_CrystalGrowth extends Chant
 	public int classificationCode(){return Ability.ACODE_CHANT|Ability.DOMAIN_ROCKCONTROL;}
 	protected int canAffectCode(){return 0;}
 	protected int canTargetCode(){return 0;}
-	protected short overrideMana(){return 50;}
+	protected int overrideMana(){return 50;}
 	public int abstractQuality(){return Ability.QUALITY_INDIFFERENT;}
 
 	public boolean invoke(MOB mob, Vector commands, Physical givenTarget, boolean auto, int asLevel)
