+++ com/planet_ink/coffee_mud/Abilities/Prayers/Prayer_Deathfinger.java
@@ -41,7 +41,7 @@ public class Prayer_Deathfinger extends Prayer
 	public int classificationCode(){return Ability.ACODE_PRAYER|Ability.DOMAIN_VEXING;}
 	public int abstractQuality(){ return Ability.QUALITY_MALICIOUS;}
 	public long flags(){return Ability.FLAG_UNHOLY;}
-	protected short overrideMana(){return 150;}
+	protected int overrideMana(){return 150;}
 
 	public boolean invoke(MOB mob, Vector commands, Physical givenTarget, boolean auto, int asLevel)
 	{
