+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_RepairingAura.java
@@ -41,7 +41,7 @@ public class Spell_RepairingAura extends Spell
 	protected int canTargetCode(){return CAN_ITEMS;}
 	public int classificationCode(){return Ability.ACODE_SPELL|Ability.DOMAIN_ABJURATION;}
 	public int abstractQuality(){ return Ability.QUALITY_INDIFFERENT;}
-	public short overrideMana(){ return 50;}
+	public int overrideMana(){ return 50;}
 	public static final int REPAIR_MAX=30;
 	public int repairDown=REPAIR_MAX;
 	public int adjustedLevel=1;
